refactor: move resources component in api package
