+++ gravitee-apim-console-webui/src/management/api/proxy/api-proxy.module.ts
@@ -21,7 +21,6 @@ import { ApiProxyResponseTemplatesModule } from './response-templates/api-proxy-
 import { ApiProxyEndpointModule } from './endpoints/api-proxy-endpoints.module';
 import { ApiProxyFailoverModule } from './failover/api-proxy-failover.module';
 import { ApiProxyHealthCheckModule } from './health-check/api-proxy-health-check.module';
-import { ApiResourcesModule } from './resources-ng/api-resources.module';
 import { ApiProxyHealthCheckDashboardModule } from './health-check-dashboard/api-proxy-health-check-dashboard.module';
 import { ApiV1PropertiesComponent } from './properties-v1/properties.component';
 import { ApiV1ResourcesComponent } from './resources-v1/resources.component';
@@ -42,7 +41,6 @@ import { ApiProxyCorsModule } from '../cors/api-proxy-cors.module';
     ApiProxyHealthCheckDashboardModule,
     ApiPropertiesModule,
     ApiDynamicPropertiesModule,
-    ApiResourcesModule,
   ],
 })
 export class ApiProxyModule {}
