+++ gravitee-apim-console-webui/src/management/api/resources-ng/api-resources.component.html
