+++ gravitee-apim-console-webui/src/management/api/resources-ng/api-resources.component.scss
@@ -1,7 +1,7 @@
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
 
-@use '../../../../scss/gio-layout' as gio-layout;
+@use '../../../scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-full-width-content-container;
@@ -13,6 +13,7 @@
 .resources {
   background-color: white;
   flex: 1 1 auto;
+  margin-top: 24px;
 
   &.save-bar-opened {
     --gv-policy-studio--pb: 84px;
