+++ gravitee-apim-console-webui/src/management/api/resources-ng/api-resources.component.spec.ts
@@ -26,10 +26,10 @@ import { ActivatedRoute } from '@angular/router';
 import { ApiResourcesComponent } from './api-resources.component';
 import { ApiResourcesModule } from './api-resources.module';
 
-import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../../shared/testing';
-import { fakeResourceListItem } from '../../../../entities/resource/resourceListItem.fixture';
-import { ApiV4, fakeApiV4 } from '../../../../entities/management-api-v2';
-import { GioTestingPermissionProvider } from '../../../../shared/components/gio-permission/gio-permission.service';
+import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../shared/testing';
+import { fakeResourceListItem } from '../../../entities/resource/resourceListItem.fixture';
+import { ApiV4, fakeApiV4 } from '../../../entities/management-api-v2';
+import { GioTestingPermissionProvider } from '../../../shared/components/gio-permission/gio-permission.service';
 
 describe('PolicyStudioResourcesComponent', () => {
   let fixture: ComponentFixture<ApiResourcesComponent>;
