+++ gravitee-apim-console-webui/src/management/api/resources-ng/api-resources.module.ts
